// libs/enums/chat.enum.ts
export enum ChatType {
	SUPPORT = 'SUPPORT',
	SELLER = 'SELLER',
}

export enum ChatStatus {
	ACTIVE = 'ACTIVE',
	CLOSED = 'CLOSED',
}

export enum MessageType {
	TEXT = 'TEXT',
	IMAGE = 'IMAGE',
	PRODUCT = 'PRODUCT',
	ORDER = 'ORDER',
	SYSTEM = 'SYSTEM',
}

export enum MessageStatus {
	SENT = 'SENT',
	READ = 'READ',
	DELETE = 'DELETE',
}